refactor(http): extract server support crate (#852)
