+++ crates/dcc-mcp-http-server/Cargo.toml
@@ -1,0 +1,38 @@
+[package]
+name = "dcc-mcp-http-server"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+readme = "README.md"
+description = "Runtime support layer for the DCC MCP HTTP server — executor, sessions, inflight requests, and notifications."
+keywords = ["mcp", "http", "server", "runtime"]
+categories = ["web-programming::http-server"]
+
+[lints]
+workspace = true
+
+[dependencies]
+# Internal crates
+dcc-mcp-http-types = { path = "../dcc-mcp-http-types" }
+dcc-mcp-jsonrpc = { path = "../dcc-mcp-jsonrpc" }
+dcc-mcp-job = { path = "../dcc-mcp-job" }
+dcc-mcp-host = { path = "../dcc-mcp-host" }
+dcc-mcp-workflow = { path = "../dcc-mcp-workflow" }
+
+# Workspace shared
+serde = { workspace = true }
+serde_json = { workspace = true }
+thiserror = { workspace = true }
+tracing = { workspace = true }
+tokio = { workspace = true }
+tokio-util = { version = "0.7", features = ["rt"] }
+dashmap = { workspace = true }
+parking_lot = { workspace = true }
+uuid = { workspace = true }
+workspace-hack = { version = "0.1", path = "../workspace-hack" }
+
+[dev-dependencies]
+tokio = { workspace = true, features = ["rt-multi-thread", "macros", "test-util"] }
