+++ crates/dcc-mcp-http/Cargo.toml
@@ -11,6 +11,7 @@ description = "MCP Streamable HTTP server for DCC applications (axum-based, thre
 [dependencies]
 # Internal crates
 dcc-mcp-http-types = { path = "../dcc-mcp-http-types" }
+dcc-mcp-http-server = { path = "../dcc-mcp-http-server" }
 dcc-mcp-actions = { path = "../dcc-mcp-actions" }
 dcc-mcp-naming = { path = "../dcc-mcp-naming" }
 dcc-mcp-protocols = { path = "../dcc-mcp-protocols" }
